feat: improve agent creation and decoupling
