+++ telegram-bot/bot/config.ts
@@ -1,5 +1,5 @@
 import logger from "bot/logger.js";
-import { kv } from "connections/redis";
+import { redis } from "connections/redis";
 import lang from "iso-language-codes";
 import { z } from "zod";
 
@@ -22,7 +22,7 @@ export const getConfigValue = async <K extends keyof typeof configSchema.shape>(
   telegramUserId: number,
   key: K,
 ): Promise<z.infer<typeof configSchema>[K]> => {
-  const config = await kv.HGETALL(`config:${telegramUserId}`);
+  const config = await redis.HGETALL(`config:${telegramUserId}`);
 
   logger.debug(config);
