+++ telegram-bot/bot/tool-data.ts
@@ -1,0 +1,5 @@
+import type { BotContext } from "bot";
+
+export type ToolData = {
+  ctx: BotContext;
+};
