+++ telegram-bot/connections/redis.ts
@@ -1,15 +1,10 @@
-import logger from "bot/logger";
 import { createClient } from "redis";
 import { getEnv } from "utils/env";
 
 const client = createClient({
-  url: `rediss://default:${getEnv("REDIS_PASSWORD")}@${getEnv(
-    "REDIS_ENDPOINT",
-  )}:6379`,
+  url: getEnv("REDIS_URL"),
 });
-
-client?.on?.("error", (error: unknown) => logger.error(error, "Redis Client Error"));
 
 await client.connect();
 
-export { client as kv };
+export { client as redis };
