+++ telegram-bot/package.json
@@ -29,7 +29,6 @@
     "@agentic/weather": "^7.6.7",
     "@ai-sdk/openai": "^1.3.22",
     "@bonfida/spl-name-service": "^3.0.11",
-    "@grammyjs/commands": "^1.0.8",
     "@grammyjs/conversations": "^2.1.0",
     "@grammyjs/parse-mode": "^2.1.2",
     "@grammyjs/runner": "^2.0.3",
@@ -63,7 +62,9 @@
     "telegramify-markdown": "^1.3.0",
     "tiktoken": "^1.0.21",
     "turndown": "^7.2.0",
-    "zod": "^3.25.56"
+    "yaml": "^2.8.0",
+    "zod": "^3.25.56",
+    "zod-to-json-schema": "^3.24.5"
   },
   "patchedDependencies": {
     "openai-zod-to-json-schema@1.0.3": "patches/openai-zod-to-json-schema@1.0.3.patch"
